backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.32
psycopg[binary]==3.2.1
mysqlclient==2.2.4
requests==2.32.3
beautifulsoup4==4.12.3
pydantic==2.9.2
python-dotenv==1.0.1
langchain-core==0.3.16
langchain-community==0.3.3
langchain-google-genai==2.0.2
